Update rule-tester and rename file
