+++ package.json
@@ -22,7 +22,7 @@
     "@types/node": "^22.7.9",
     "@typescript-eslint/eslint-plugin": "^7.8.0",
     "@typescript-eslint/parser": "^7.8.0",
-    "@typescript-eslint/rule-tester": "7.5.0",
+    "@typescript-eslint/rule-tester": "^8.13.0",
     "@typescript-eslint/types": "7.5.0",
     "@typescript-eslint/utils": "7.5.0",
     "eslint": "8.56.0",
